# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API WitelonBank",
        "description": "Dokumentacja API dla systemu bankowego WitelonBank. Pamiętaj o autoryzacji (przycisk 'Authorize' na górze) dla chronionych endpointów.",
        "contact": {
            "email": "support@witelonbank.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Główny serwer API"
        }
    ],
    "paths": {
...
        "/api/przelewy": {
            "post": {
                "tags": [
                    "Przelewy"
                ],
                "summary": "Tworzy nowy przelew",
                "operationId": "7f3ada558b127ed64b63efb9ed499b86",
                "requestBody": {
                    "description": "Dane przelewu",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_konta_nadawcy",
                                    "nr_konta_odbiorcy",
                                    "nazwa_odbiorcy",
                                    "tytul",
                                    "kwota",
                                    "waluta_przelewu"
                                ],
                                "properties": {
                                    "id_konta_nadawcy": {
                                        "description": "ID konta, z którego wysyłany jest przelew",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nr_konta_odbiorcy": {
                                        "description": "Numer konta odbiorcy",
                                        "type": "string",
                                        "example": "PL98765432109876543210987654"
                                    },
                                    "nazwa_odbiorcy": {
                                        "description": "Nazwa odbiorcy",
                                        "type": "string",
                                        "example": "Anna Nowak"
                                    },
                                    "adres_odbiorcy_linia1": {
                                        "type": "string",
                                        "example": "ul. Słoneczna 10",
                                        "nullable": true
                                    },
                                    "adres_odbiorcy_linia2": {
                                        "type": "string",
                                        "example": "00-123 Warszawa",
                                        "nullable": true
                                    },
                                    "tytul": {
                                        "type": "string",
                                        "example": "Za zakupy"
                                    },
                                    "kwota": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150.99
                                    },
                                    "waluta_przelewu": {
                                        "description": "Waluta przelewu (np. PLN, EUR)",
                                        "type": "string",
                                        "example": "PLN"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Przelew został pomyślnie utworzony (zlecony)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PrzelewResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Błąd danych (np. niewystarczające środki, konto zablokowane)"
                    },
                    "401": {
                        "description": "Nieautoryzowany"
                    },
                    "403": {
                        "description": "Brak uprawnień do konta nadawcy"
                    },
                    "404": {
                        "description": "Konto nadawcy nie znalezione"
                    },
                    "422": {
                        "description": "Błąd walidacji"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/konta/{idKonta}/przelewy": {
            "get": {
                "tags": [
                    "Przelewy"
                ],
                "summary": "Pobiera historię przelewów dla danego konta",
                "operationId": "c0ff505db5be916e7fb4cfe0ee7c88c0",
                "parameters": [
                    {
                        "name": "idKonta",
                        "in": "path",
                        "description": "ID konta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "typ",
                        "in": "query",
                        "description": "Filtruj wg typu: 'wychodzace', 'przychodzace'",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "wychodzace",
                                "przychodzace"
                            ]
                        }
                    },
                    {
                        "name": "strona",
                        "in": "query",
                        "description": "Numer strony dla paginacji",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "na_strone",
                        "in": "query",
                        "description": "Liczba wyników na stronę",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista przelewów",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PrzelewResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Nieautoryzowany"
                    },
                    "403": {
                        "description": "Brak uprawnień do tego konta"
                    },
                    "404": {
                        "description": "Konto nie znalezione"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/konta/{konto}/przelewy/export": {
            "get": {
                "tags": [
                    "Przelewy"
                ],
                "summary": "Eksportuje historię transakcji dla danego konta do pliku CSV",
                "operationId": "9f0f682349fa14dac6a04b04b7121490",
                "parameters": [
                    {
                        "name": "konto",
                        "in": "path",
                        "description": "ID konta, dla którego generowany jest raport",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "data_od",
                        "in": "query",
                        "description": "Data początkowa zakresu raportu (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "data_do",
                        "in": "query",
                        "description": "Data końcowa zakresu raportu (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "typ",
                        "in": "query",
                        "description": "Filtruj wg typu przelewu ('wychodzacy' lub 'przychodzacy')",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "wychodzacy",
                                "przychodzacy"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plik CSV z historią transakcji",
                        "content": {
                            "text/csv": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Nieautoryzowany"
                    },
                    "403": {
                        "description": "Brak uprawnień do tego konta"
                    },
                    "404": {
                        "description": "Konto nie znalezione"
                    },
                    "422": {
                        "description": "Błąd walidacji parametrów"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Wprowadź token JWT z prefiksem Bearer (np. 'Bearer eyJhbGciOiJIUzI1NiI...). Token uzyskasz po zalogowaniu.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}